Add attributes to dist.ini
Adds author and copyright_holder attributes to dist.ini. If these are
missing when running "dzil build", Dist::Zilla::Plugin::PodWeaver will
use the values in ~/dzil/config.ini when creating the pod document.

This Pod document will then be converted to a README.md file in the root
directory by the plugin Dist::Zilla::Plugin::ReadmeAnyFromPod. Since
this README.md is part of the github repository's tracked files it could
be inconvenient for other users than the author of this module to submit
changes to README.md. They would then have to first revert the changes
to the AUTHOR and "COPYRIGHT AND LICENSE" sections in README.md made by
"dzil build" before changing the file (as part of a potential pull request).

Alternatively, if they do not run "dzil build" the reasoning above does
not apply. But I think running "dzil build" or "dzil test" is something one
often will do before submitting a PR.